update forklift to use single utility for starting and stopping data readers
smartcitiesdata/smartcitiesdata#430
