+++ apps/forklift/config/dev.exs
@@ -9,7 +9,7 @@ config :forklift,
 config :prestige, base_url: "http://127.0.0.1:8080"
 
 config :redix,
-       args: [host: "localhost"]
+  args: [host: "localhost"]
 
 config :forklift, :brook,
   instance: :forklift,
